pull useAllPlansRLO hook out of nav controller
