+++ src/features/Navigation/NavigationController.tsx
@@ -8,17 +8,13 @@ import {
 } from "features/Navigation/components/Navigation.elements";
 import { FlexBox } from "global/components/FlexBox";
 import useFluxStore from "data/useFluxStore";
-import { ripLoadObject } from "util/ripLoadObject";
-import planStore from "features/Planner/data/planStore";
 import useIsDevMode from "data/useIsDevMode";
 import { useIsMobile } from "providers/IsMobile";
 import { MobileNav } from "features/Navigation/components/MobileNav";
 import { DesktopNav } from "features/Navigation/components/DesktopNav";
 import { useHistory } from "react-router-dom";
-import { useLogoutHandler, useProfileLO } from "providers/Profile";
+import { useLogoutHandler } from "providers/Profile";
 import RouteStore from "../../data/RouteStore";
-import friendStore from "../../data/FriendStore";
-import { zippedComparator } from "../../util/comparators";
 import Dispatcher from "../../data/dispatcher";
 import ShoppingActions from "../../data/ShoppingActions";
 import PlanActions from "../Planner/data/PlanActions";
@@ -51,7 +47,6 @@ export const NavigationController: React.FC<NavigationControllerProps> = ({
     const expanded = !useIsNavCollapsed();
     const isMobile = useIsMobile();
     const devMode = useIsDevMode();
-    const profileRLO = ripLoadObject(useProfileLO());
     const history = useHistory();
     const [selected, setSelected] = React.useState("");
 
@@ -95,47 +90,6 @@ export const NavigationController: React.FC<NavigationControllerProps> = ({
         : selectPlan;
     const handleOpenPlan = shopView ? toggleShoppingPlan : openPlan;
 
-    const getPlans = useFluxStore(
-        () => {
-            const allPlans = ripLoadObject(
-                planStore.getPlansLO().map((plans) => {
-                    const myId = profileRLO.data && profileRLO.data.id;
-                    const orderComponentsById = plans.reduce((byId, p) => {
-                        let ownerId =
-                            (p.acl ? p.acl.ownerId : undefined) ||
-                            Number.MAX_SAFE_INTEGER;
-                        let ownerName = "";
-                        if (ownerId === myId) {
-                            ownerId = -1;
-                        } else {
-                            const rlo = ripLoadObject(
-                                friendStore.getFriendLO(ownerId),
-                            );
-                            if (rlo.data) {
-                                // eslint-disable-next-line @typescript-eslint/no-extra-non-null-assertion
-                                ownerName = rlo.data.name!!; // eslint-disable-line @typescript-eslint/no-non-null-assertion
-                            }
-                        }
-                        byId[p.id] = [ownerId, ownerName, p.name.toLowerCase()];
-                        return byId;
-                    }, {});
-                    return plans
-                        .slice()
-                        .sort((a, b) =>
-                            zippedComparator(
-                                orderComponentsById[a.id],
-                                orderComponentsById[b.id],
-                            ),
-                        );
-                }),
-            );
-            return { allPlans };
-        },
-        [planStore, friendStore],
-        [profileRLO.data],
-    );
-    const { data: navPlanItems } = getPlans.allPlans;
-
     if (!authenticated) {
         return isMobile ? (
             <MainMobile>{children}</MainMobile>
@@ -168,7 +122,6 @@ export const NavigationController: React.FC<NavigationControllerProps> = ({
                 onOpenPlan={handleOpenPlan}
                 onExpand={handleExpand}
                 devMode={devMode}
-                planItems={navPlanItems}
             />
             <MainDesktop open={expanded}>{children}</MainDesktop>
         </FlexBox>
